+++ packages/mod-types/src/index.ts
@@ -2,6 +2,7 @@ export * from "./assets";
 export * from "./translation";
 export * from "./imageMapping";
 export * from "./activity";
+export * from "./logger";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type FuncWork<T extends any[] = []> = (...args: T) => void;