+++ packages/asset-manager/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sugarch/bc-asset-manager
 
+## 0.2.14
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @sugarch/bc-mod-types@0.5.0
+    -   @sugarch/bc-mod-hook-manager@0.2.10
+
 ## 0.2.13
 
 ### Patch Changes