+++ packages/image-mapping/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sugarch/bc-image-mapping
 
+## 1.0.4
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @sugarch/bc-mod-types@0.5.0
+    -   @sugarch/bc-mod-hook-manager@0.2.10
+
 ## 1.0.3
 
 ### Patch Changes